Remove GEMINI_SETUP.md file and update project name from "shopyy" to "ayala-cakes" in package.json and package-lock.json. Refactor import paths in multiple components to use absolute paths. Add new documentation files for Critical Rules and SEO & AIO Scoring. Introduce legacy components and scripts for product management and order settings.
