+++ Components/ProductItem/index.jsx
@@ -5,7 +5,7 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import axios from 'axios';
-import useStore from '../../useStore';
+import useStore from '@/store/useStore';
 
 
 export default function ProductItem({ product, hideAdminActions = false }) {
