--- GEMINI_SETUP.md
@@ -1,76 +1,0 @@
-# הגדרת Gemini API לג'ינרוט SEO
-
-## שלבי ההתקנה
-
-### 1. קבלת API Key מ-Google
-1. היכנס לאתר [Google AI Studio](https://makersuite.google.com/app/apikey)
-2. צור API Key חדש
-3. העתק את המפתח
-
-### 2. הוספת המפתח לפרויקט
-צור קובץ `.env.local` בתיקיית הפרויקט והוסף:
-
-```env
-GEMINI_API_KEY=your_api_key_here
-# אופציונלי — ברירת מחדל: gemini-3.5-flash
-GEMINI_MODEL=gemini-3.5-flash
-```
-
-### 3. אימות ההתקנה
-אחרי הוספת המפתח, אתה יכול להשתמש בכפתור "ייצר תוכן SEO באמצעות AI" בטפסי הוספת מוצרים ופוסטים.
-
-## תכונות שנוספו
-
-### למוצרים:
-- ג'ינרוט SEO Title
-- ג'ינרוט Meta Description
-- ג'ינרוט מילות מפתח
-- ג'ינרוט Slug באנגלית
-- ג'ינרוט Alt Text לתמונות
-- ג'ינרוט תגיות
-- ג'ינרוט Structured Data
-
-### לפוסטים:
-- ג'ינרוט SEO Title
-- ג'ינרוט Meta Description
-- ג'ינרוט מילות מפתח
-- ג'ינרוט Slug באנגלית
-- ג'ינרוט קריאה לפעולה
-- חישוב זמן קריאה אוטומטי
-- ברירת מחדל: מחבר = "אילה אברהם"
-
-## איך זה עובד?
-
-1. מלא את הפרטים הבסיסיים של המוצר/פוסט
-2. לחץ על כפתור "ייצר תוכן SEO באמצעות AI"
-3. הAI יצור תוכן SEO מתקדם בעברית
-4. תוכל לערוך את התוכן שנוצר לפני שמירה
-5. התוכן נשמר באופן אוטומטי עם המוצר/פוסט
-
-## פתרון בעיות נפוצות
-
-### שגיאת "models/gemini-... is not found" (404)
-Google מפסיקה מודלים ישנים. אם מופיעה שגיאה על `gemini-2.5-flash` או מודל אחר — עדכן ב-`.env.local`:
-
-```env
-GEMINI_MODEL=gemini-3.5-flash
-```
-
-או מחק את השורה `GEMINI_MODEL` כדי להשתמש בברירת המחדל של המערכת.
-
-### שגיאת API Key
-- וודא שהמפתח שלך נכון ופעיל
-- בדוק שקובץ `.env.local` נמצא בתיקיית הפרויקט הראשית
-- וודא שלא יש רווחים נוספים במפתח
-
-### שגיאת Quota
-אם המפתח שלך הגיע לגבול השימוש היומי, תצטרך לחכות או לשדרג את התוכנית שלך ב-Google AI Studio.
-
-## הערות חשובות
-
-- הג'ינרוט דורש חיבור לאינטרנט
-- הSlug נוצר באנגלית לנגישות SEO טובה יותר
-- כל התוכן האחר נוצר בעברית
-- ניתן לערוך את התוכן שנוצר לפני השמירה
-- המערכת משתמשת במודל `gemini-3.5-flash` (ניתן לשנות ב-`GEMINI_MODEL`)
-- מחבר הפוסטים מוגדר אוטומטית ל"אילה אברהם"
