+++ legacy/README.md
@@ -1,0 +1,5 @@
+# Legacy (unused)
+
+Parked code that is **not imported by the live app**. Kept for reference / possible revive — do not import from here in new work.
+
+Restore by moving a folder/file back into `Components/`, `server/`, `utils/`, etc., then wiring imports.
