+++ legacy/scripts/migrate-products-slug.js
