+++ legacy/server/middleware/authMiddleware.js
