+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "shopyy",
+  "name": "ayala-cakes",
   "version": "0.1.0",
   "private": true,
   "scripts": {
